requirements.txt
Pillow==11.0.0
python-telegram-bot==21.7
PyYAML==6.0.2
Requests==2.32.3
websockets==12.0
anthropic